[ty] Extract project rediscovery
